+++ HISTORY.md
@@ -23,6 +23,7 @@
 * byob, objects: new feature: hide custom helper blocks in palette
 * German translation update (hiding helper blocks in the palette)
 * gui: unified palette: offer the currently visible category when letting the user make a block
+* scenes, objects, gui: made unified palette the default for dev
 
 ### 2021-07-05
 * gui: unified palette: indicate "selected" category in selector buttons